+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Implement rfc show, adr show, work show, clause show commands (WI-2026-02-07-002)
+- Support -o json for structured output (WI-2026-02-07-002)
+- Update describe output with show commands (WI-2026-02-07-002)
+
 ## [0.3.1] - 2026-02-07
 
 ### Added